Extract common parts from "prod" config to "base"
