+++ package.json
@@ -24,6 +24,7 @@
     "@types/react": "16.9.17",
     "@types/react-dom": "16.9.4",
     "@types/webpack": "4.41.1",
+    "@types/webpack-merge": "4.1.5",
     "@typescript-eslint/eslint-plugin": "2.15.0",
     "@typescript-eslint/parser": "2.15.0",
     "clean-webpack-plugin": "3.0.0",
@@ -41,7 +42,8 @@
     "ts-node": "8.5.4",
     "typescript": "3.7.4",
     "webpack": "4.41.5",
-    "webpack-cli": "3.3.10"
+    "webpack-cli": "3.3.10",
+    "webpack-merge": "4.2.2"
   },
   "dependencies": {
     "react": "16.12.0",
